dividing mongo sherpa to mongo cluster care and seacat auth care
