+++ Site/ASAB Maestro/Descriptors/seacat-auth.yaml
@@ -13,6 +13,18 @@ descriptor:
     - "{{SITE}}/{{INSTANCE_ID}}/conf:/conf:ro"
     - "{{FAST_STORAGE}}/{{INSTANCE_ID}}/data:/data"
 
+sherpas:
+# Sherpas containers: akin to their namesake mountain guides, these containers provide essential support and guidance throughout the application's lifecycle.
+# provide a name to your sherpa and a descriptor for its very own container.
+  mongo:
+    image: library/mongo
+    entrypoint: ["mongosh", "--nodb", "--file", "/script/mongo-insert.js"]
+    command: ["echo", "DONE"]
+    volumes:
+    - "{{SITE}}/{{INSTANCE_ID}}/script:/script"
+    environment:
+      MONGO_HOSTNAMES: "{{MONGO_HOSTNAMES}}"
+
 files:
   # By default, SMTP configuration is mocked - e-mails are printed into log of seacat-auth
   # Configuration of SMTP server should be done in UI > Maintenance > Configuration > General
@@ -36,6 +48,9 @@ files:
 
       [seacatauth:communication:email:smtp]
       host=<mocked>
+
+  - "script/mongo-insert.js"
+  - "script/to_upload/"
 
 api:
   public:
